Finish moving tests out of migration_tests.rs
This lets us simplify the migration_engine_tests API by a lot.
