+++ migration-engine/migration-engine-tests/Cargo.toml
@@ -27,7 +27,6 @@ bigdecimal = "0.2"
 chrono = "0.4.15"
 enumflags2 = "0.7"
 indoc = "1.0.3"
-once_cell = "1.5"
 pretty_assertions = "0.6"
 serde = "1"
 serde_json = { version = "1.0", features = ["float_roundtrip"] }
@@ -38,6 +37,7 @@ tracing-futures = "0.2.1"
 url = "2.1.1"
 connection-string = "0.1.13"
 expect-test = "1.1.0"
+once_cell = "1.8.0"
 
 [dependencies.quaint]
 git = "https://github.com/prisma/quaint"
