+++ migration-engine/migration-engine-tests/src/commands.rs
@@ -1,0 +1,21 @@
+mod apply_migrations;
+mod create_migration;
+mod dev_diagnostic;
+mod diagnose_migration_history;
+mod evaluate_data_loss;
+mod list_migration_directories;
+mod mark_migration_applied;
+mod mark_migration_rolled_back;
+mod reset;
+mod schema_push;
+
+pub(crate) use apply_migrations::*;
+pub(crate) use create_migration::*;
+pub(crate) use dev_diagnostic::*;
+pub(crate) use diagnose_migration_history::*;
+pub(crate) use evaluate_data_loss::*;
+pub(crate) use list_migration_directories::*;
+pub(crate) use mark_migration_applied::*;
+pub(crate) use mark_migration_rolled_back::*;
+pub(crate) use reset::*;
+pub(crate) use schema_push::*;
