--- migration-engine/migration-engine-tests/src/sql.rs
@@ -1,7 +1,0 @@
-mod quaint_result_set_ext;
-
-pub use super::{assertions::*, test_api::*};
-pub use quaint::prelude::Queryable;
-pub use quaint_result_set_ext::*;
-pub use test_macros::test_connector;
-pub use test_setup::{BitFlags, Capabilities, Tags};
