+++ migration-engine/migration-engine-tests/tests/apply_migrations/mod.rs
@@ -1,5 +1,5 @@
 use indoc::formatdoc;
-use migration_engine_tests::{sync_test_api::*, MigrationsAssertions};
+use migration_engine_tests::sync_test_api::*;
 use pretty_assertions::assert_eq;
 use user_facing_errors::{migration_engine::ApplyMigrationError, UserFacingError};
